All the entities have been moved to new pack system

+++ main/java/minecrafttransportsimulator/systems/pack/PackCollisionBox.java
@@ -3,8 +3,8 @@ package minecrafttransportsimulator.systems.pack;
 
 public class PackCollisionBox {
 
-    public int[] pos;
-    public int width;
-    public int height;
+    public float[] pos;
+    public float width;
+    public float height;
 
 }
